ALPHA-211 : Patch[96223] : 9.5c_update_VQE_library_v02 (#42)
* ALPHA-208_A: Patch[96029] : v9.5c_AURU_Sensor_BringUp_1361_(V01)

* ALPHA-209_B : v9.5c_fix_printf_assert_in_critical_section(96148)

* ALPHA-201: Patch[96206]: v9.5c_patch_support_tuning_tool_(v01)

* ALPHA-211: Patch[96223] : 9.5c_update_VQE_library_v02

+++ component/audio/3rdparty/AEC/ASP.h
@@ -26,7 +26,7 @@ typedef enum {
 //Setting for ASP
 typedef struct CTNS_cfg_s {
 	int16_t NS_EN;
-	int NSLevel;
+	int16_t NSLevel;
 
 	int16_t Reserve1;
 	int16_t Reserve2;
